requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
automat==24.8.1
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
mako==1.3.9
markupsafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.39
starlette==0.46.1
tornado==6.4.2
trio==0.29.0
twisted==24.11.0
typing-extensions==4.12.2
uvicorn==0.34.0
zope-interface==7.2
